Inline TaskType editor on Settings, matching Resource editing pattern
Replace the TaskType dialog with an inline edit row in the Task Types
list, mirroring the EditingRow pattern used in ResourcesTable on Task
pages. Editing a row swaps it for the form in place; "New Task Type"
spawns a draft row at the top of the list. Edit/Delete on other rows
are disabled while one is being edited.
